requirements.txt
PlexAPI
pre-commit
pytest
requests
python-dotenv
